backend/pyproject.toml
[project]
name = "docguard"
version = "0.1.0"
requires-python = ">=3.12"

dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.32.0",
    "mangum==0.19.0",                        # ASGI adapter for AWS Lambda
    "pydantic==2.9.2",
    "pydantic-settings==2.6.1",
    "sqlalchemy[asyncio]==2.0.36",
    "alembic==1.14.0",
    "asyncpg==0.30.0",
    "httpx==0.27.2",
    "openai==1.54.4",                        # OpenAI SDK → OpenRouter
    "PyGithub==2.4.0",
    "python-jose[cryptography]==3.3.0",      # JWT verification
    "markdown-it-py==3.0.0",
    "python-multipart==0.0.12",
    "structlog==24.4.0",
]

[project.optional-dependencies]
dev = [
    "ruff==0.7.4",
    "mypy==1.13.0",
    "pytest==8.3.3",
    "pytest-asyncio==0.24.0",
    "pytest-cov==6.0.0",
    "respx==0.21.1",
    "faker==33.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501", "B008"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=src --cov-report=term-missing"
